Refactored multiple calls into one transactional confirm-endpoint and splitted the Orders.vue.
