--- app/Http/Requests/Order/UpdateReceiptProductPriceRequest.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Order;
-
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rule;
-
-class UpdateReceiptProductPriceRequest extends FormRequest
-{
-    public function authorize(): bool
-    {
-        return auth()->check();
-    }
-
-    public function rules(): array
-    {
-        return [
-            'product_id' => [
-                'required',
-                'integer',
-                Rule::exists('products', 'id')->where('company_id', auth()->user()->company->id),
-            ],
-            'price' => ['required', 'numeric', 'min:0'],
-        ];
-    }
-}
