+++ resources/js/Components/Products/productCard.vue
@@ -46,7 +46,7 @@ export default {
         <div class="mt-3 sm:mt-0 ">
             <div class="text-sm font-medium text-gray-900">{{ product.name }}</div>
             <div class="mt-1">
-                <span v-if="product.variable_price" class="text-xs text-gray-400">Price per kg</span>
+                <span v-if="product.variable_price" class="text-xs text-gray-400">Indicative price</span>
             </div>
         </div>
         <div class="card-actions justify-end mt-2">
@@ -55,7 +55,7 @@ export default {
                 class="btn btn-sm btn-success modal-button">
                 Order
                 <div class="badge bg-white text-success font-bold border-0 ml-3">
-                    {{ formatMoney(product.price) }}{{ product.variable_price ? '/kg' : '' }}
+                    {{ product.variable_price ? '≈ ' : '' }}{{ formatMoney(product.price) }}
                 </div>
             </label>
             <input type="checkbox" :id="'option-modal-'+product.id" class="modal-toggle" />
@@ -64,11 +64,11 @@ export default {
                 <label class="modal-box relative" for="">
                     <h3 class="text-lg font-bold">{{product.name}}
                         <span class="ml-2 badge badge-warning badge-outline p-1">
-                            {{ product.variable_price ? formatMoney(product.price) + '/kg' : formatMoney(product.price) }}
+                            {{ product.variable_price ? '≈ ' + formatMoney(product.price) : formatMoney(product.price) }}
                         </span>
                     </h3>
                     <div v-if="product.variable_price" class="mt-3 p-3 bg-amber-50 rounded-lg text-sm text-amber-700">
-                        The price is determined by the runner based on the weight ({{ formatMoney(product.price) }}/kg).
+                        This price is indicative ({{ formatMoney(product.price) }}) — the final price is filled in after pickup.
                     </div>
                     <div v-if="options.length >=0">
                         <p v-if="options.length" class="py-4">Choose your options</p>
