+++ routes/api.php
@@ -6,19 +6,16 @@ use App\Http\Controllers\Order\AddProductController;
 use App\Http\Controllers\Order\AnalyzeReceiptController;
 use App\Http\Controllers\Order\AssignToMeController;
 use App\Http\Controllers\Order\CheckNewStoreController;
-use App\Http\Controllers\Order\ConfirmDeliveryPricesController;
+use App\Http\Controllers\Order\ConfirmDeliveryController;
 use App\Http\Controllers\Order\DepartAsRunnerController;
 use App\Http\Controllers\Order\GetDoneOrdersController;
 use App\Http\Controllers\Order\GetOrdersByDateController;
 use App\Http\Controllers\Order\GetOrdersController;
 use App\Http\Controllers\Order\GetSelectedRunnerController;
 use App\Http\Controllers\Order\GetSimulatedRunnerController;
-use App\Http\Controllers\Order\MarkAsDeliveredController;
 use App\Http\Controllers\Order\RemoveProductController;
-use App\Http\Controllers\Order\StoreReceiptProductController;
 use App\Http\Controllers\Order\StoreReceiptStoreController;
-use App\Http\Controllers\Order\UpdateReceiptProductPriceController;
-use App\Http\Controllers\Order\UpdateWeightController;
+use App\Http\Controllers\Order\UpdateOrderPriceController;
 use App\Http\Controllers\PayoutController;
 use App\Http\Controllers\Store\AddStoreController;
 use App\Http\Controllers\Store\DeleteProductController;
@@ -41,14 +38,11 @@ Route::post('/users', [UserController::class, 'index'])->name('users.index');
 Route::post('/orders', GetOrdersController::class)->name('orders.index');
 Route::post('/done-orders', GetDoneOrdersController::class)->name('orders.done');
 Route::post('/assign-to-me', AssignToMeController::class)->name('order.assign-to-me');
-Route::patch('/order/deliver', MarkAsDeliveredController::class)->name('order.deliver');
 Route::patch('/order/depart', DepartAsRunnerController::class)->name('order.depart');
-Route::patch('/order/weight', UpdateWeightController::class)->name('order.weight');
+Route::patch('/order/price', UpdateOrderPriceController::class)->name('order.price');
 Route::post('/order/receipt', AnalyzeReceiptController::class)->name('order.receipt.analyze');
 Route::post('/order/receipt/store', StoreReceiptStoreController::class)->name('order.receipt.store');
-Route::post('/order/receipt/product', StoreReceiptProductController::class)->name('order.receipt.product');
-Route::patch('/order/receipt/price', UpdateReceiptProductPriceController::class)->name('order.receipt.price');
-Route::patch('/order/prices', ConfirmDeliveryPricesController::class)->name('order.prices');
+Route::post('/order/confirm-delivery', ConfirmDeliveryController::class)->name('order.confirm-delivery');
 Route::post('/selected-runner', GetSelectedRunnerController::class)->name('order.selected-runner');
 Route::post('/simulated-runner', GetSimulatedRunnerController::class)->name('order.simulated-runner');
